api/requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
dj-database-url==3.1.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dlib==19.24.9
face-recognition==1.3.0
face_recognition_models==0.3.0
h11==0.16.0
numpy==2.2.6
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.42.0
whitenoise==6.12.0
